Continue refactor. Simplify __init__, and move logic to reasonably named files
